+++ CHANGELOG.md
@@ -13,6 +13,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 * Fixed shape key indexing for wimdo exports.
 * Fixed an issue where camdo models would not correctly import all bones.
 * Fixed an issue where some camdo models would not import due to texture loading errors.
+* Fixed an issue where enabling speff mesh export would incorrectly genrate additional meshes for Xenoblade 1 DE and Xenoblade 2.
 
 ## 0.8.1 - 2024-06-07
 ### Fixed